Move HelpArgumentAction to helper package
